+++ example/a.ts
@@ -5,5 +5,6 @@ const counter = (app: KingWorld) => app.state('nested-counter', 1)
 new KingWorld()
 	.state('counter', 1)
 	.use(counter)
-	.use(a => a)
-	.get('/', ({ store }) => store)
+	.get('/:id', ({ store, params }) => store, {
+		beforeHandle: ({ params }) => {}
+	})